Moved semisafe out of src/composition
